Added my Shiva team. Fixed crash on overwriting a deleted team.

+++ app/src/main/java/com/example/anthony/damagecalculator/Fragments/TeamOverviewFragment.java
@@ -218,14 +218,14 @@ public class TeamOverviewFragment extends AbstractFragment {
 
     private void setTeamStats(){
         int teamHp = 0;
-        int teamRcv = 0;
+        double teamRcv = 0;
         Log.d("Team Overview", "Leadskill is: " + team.getLeadSkill());
         for(int i = 0; i < team.getMonsters().size(); i++){
             teamHp += DamageCalculationUtil.monsterHpCalc(team.getMonsters(i), team);
             teamRcv += DamageCalculationUtil.monsterRcvCalc(team.getMonsters(i), team);
         }
         teamHpValue.setText(String.valueOf(teamHp));
-        teamRcvValue.setText(String.valueOf(teamRcv));
+        teamRcvValue.setText(String.valueOf((int)teamRcv));
     }
 
     @Override
